+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1 - TBD
 
 - Support conversion of HTML comments
+- Add `unicode-subst` config option for Unicode character substitution.
 
 ## 0.1.0 - 2023-10-21